Module logos::callback

Types and traits used within lexer callbacks.

When lexing, it’s very common to want to do something whenever you encounter a certain token - you might want to convert a string slice into an integer for a hypothetical Number token, or perhaps want to process escape sequences for a hypothetical StringLiteral token. Logos allows you to perform this kind of processing using callbacks, which are attached to token variants and are called by the lexer when a token matches.

Callbacks take one argument (a mutable reference to a lexer) and can return any value implementing the CallbackResult trait. This trait is implemented for a few types already, and their behaviour is listed below.

For unit variants (variants like Token::Variant, which don’t contain data):

Type	Effect

bool	If true, emits Token::Unit. Otherwise, emits a generic “unknown token” error.

Skip	Skips the matched token

Filter<()>	If Filter::Accept(()), emits Token::Unit. Otherwise, skips the matched token.

Option<()>	If Some(()), emits Token::Unit. Otherwise, emits a generic “unknown token” error.

Result<(), E>	If Ok(()), emits Token::Unit. If Err(E), emits the contained error value.

For value variants (variants like Token::Value(C), which contain one piece of data):

Type	Effect

Filter<C>	If Filter::Accept(C), creates and emits Token::Value(C). Otherwise, skips the matched token.

Option<()>	If Some(C), emits Token::Value(C). Otherwise, emits a generic “unknown token” error.

Result<C, E>	If Ok(C), creates and emits Token::Value(C). If Err(E), emits the contained error value.

The Output type can also be returned from callbacks, and how it behaves depends on the data inside:

If Output::Skip, skips the matched token.

If Output::Construct(()), emits Token::Unit.

If Output::Construct(C), creates and emits Token::Value(C).

If Output::Emit(T), emits the token T as-is.

If Output::Error(E), emits the error value.

In order to support more sophisticated lexing strategies, Logos also allows callbacks to manipulate the lexer itself. The Lexer documentation contains more details, but you’re most likely interested in Lexer::remainder and Lexer::bump.

Enums

Output

Represents actions the lexer can take.

Traits

CallbackResult

Types that can be returned from lexer callbacks.