Trait logos::callback::CallbackResult

pub trait CallbackResult<'s, C, T> where
    T: Logos<'s>, {
    fn construct(self, lex: &Lexer<'s, T>) -> Output<C, T, T::Error>;
}

Types that can be returned from lexer callbacks.

This trait is implemented for many types already, and there probably isn’t much reason to implement it yourself. See the table in the module-level documentation for an overview.

Required methods

fn construct(self, lex: &Lexer<'s, T>) -> Output<C, T, T::Error>

Construct an Output value using self, instructing the lexer how to proceed. See Output’s documentation for more information.