Module logos::error

Types and traits for error handling in Logos.

Unfortunately, the world is not sunshine and rainbows, so - at some point - your lexer will receive invalid input. What “invalid” looks like for you depends on circumstance, but regardless of what you’re doing it’s important to be able to accurately report that something went wrong.

As part of that reporting process, you might also want to include some details about what happened, like the specific error’s cause, or where you found the error. All of this is information that can be used later down the track to help somebody (maybe even you!) figure out what went wrong.

By default, Logos uses the UnknownToken type to represent the standard fare lexing error: an unknown character or sequence of characters was found. For most people, this information is perfectly sufficient for their purposes, and anything else would be overkill. However, sometimes that information isn’t enough, and you’d like to convey a more specific cause or provide more information along with the error. To that end, Logos allows you to use your own error type when lexing, and additionally allows you to control how an error is constructed when a lexer encounters an error.

See Error’s documentation if you’d like to implement it for your own type. Otherwise, you may be interested in the Logos trait’s documentation, which covers how to use a type implementing Error with Logos.

Structs

UnknownToken

The primary error case when lexing, and the default error type in Logos.

Traits

Error

A trait for representing errors that occur during lexing.