Trait logos::internal::LexerInternal

pub trait LexerInternal<'source> {
    type Token;

    // Required methods
    fn read<T: Chunk<'source>>(&self) -> Option<T>;
    fn read_at<T: Chunk<'source>>(&self, n: usize) -> Option<T>;
    unsafe fn read_unchecked<T: Chunk<'source>>(&self, n: usize) -> T;
    fn test<T: Chunk<'source>, F: FnOnce(T) -> bool>(&self, test: F) -> bool;
    fn test_at<T: Chunk<'source>, F: FnOnce(T) -> bool>(
        &self,
        n: usize,
        test: F
    ) -> bool;
    fn bump_unchecked(&mut self, size: usize);
    fn trivia(&mut self);
    fn error(&mut self);
    fn end(&mut self);
    fn set(&mut self, token: Self::Token);
}

Trait used by the functions contained in the Lexicon.

WARNING!

This trait, and it’s methods, are not meant to be used outside of the code produced by #[derive(Logos)] macro.

Required Associated Types

type Token

Required Methods

fn read<T: Chunk<'source>>(&self) -> Option<T>

Read a chunk at current position.

fn read_at<T: Chunk<'source>>(&self, n: usize) -> Option<T>

Read a chunk at current position, offset by n.

unsafe fn read_unchecked<T: Chunk<'source>>(&self, n: usize) -> T

Unchecked read a chunk at current position, offset by n.

fn test<T: Chunk<'source>, F: FnOnce(T) -> bool>(&self, test: F) -> bool

Test a chunk at current position with a closure.

fn test_at<T: Chunk<'source>, F: FnOnce(T) -> bool>( &self, n: usize, test: F ) -> bool

Test a chunk at current position offset by n with a closure.

fn bump_unchecked(&mut self, size: usize)

Bump the position by size.

fn trivia(&mut self)

Reset token_start to token_end.

fn error(&mut self)

Set the current token to appropriate #[error] variant. Guarantee that token_end is at char boundary for &str.

fn end(&mut self)

fn set(&mut self, token: Self::Token)