Module logos::iter

Tools for working with lexers as iterators.

Rust’s standard library already contains a rich variety of functionality geared around Iterators, but a lot of it doesn’t fit in nicely with the rest of Logos - for example, the Iterator returned by Iterator::peekable doesn’t give you any access to the underlying iterator, so using peekable with a Lexer becomes very frustrating.

This is quite the pain point, and unfortunately it’s a bit hard to come up with a good solution. In an attempt to address this, Logos provides the LexerExt trait and a few special iterator types. These work together to allow you to access the underlying Lexer, even when it’s wrapped inside another iterator.

use logos::{Logos, LexerExt};

#[derive(Logos, Debug, PartialEq)]
enum Token {
    #[regex(r"\s", logos::skip)]
    Whitespace,

    #[regex(r"[1-9]+[0-9]*")]
    Integer,
}

let mut lexer = Token::lexer("1000 900 2 48").lookahead();
// We can use the iterator that `lookahead` returns as usual
let peeked = lexer.peek();
// ... and we can access information from the `Lexer` too
let current = lexer.slice();

Unfortunately, this approach comes with its own limitations. Most notably, you lose the ability to access the underlying Lexer after using an iterator adaptor from the standard library, since those types don’t implement LexerExt

Structs

BoxedLexer

A boxed and type-erased lexer.

Lookahead

An iterator with a peek() method that can look into the future.

MapWithLexer

An iterator that maps each value to another, making use of the lexer in the process.