logos
Module lexer

Structs

Lexer

Lexer is the main struct of the crate that allows you to read through a Source and produce tokens for enums implementing the Logos trait.

SpannedIter

Iterator that pairs tokens with their position in the source.

Type Aliases

Span

Byte range in the source.