logos
Module source

This module contains a bunch of traits necessary for processing byte strings.

Most notable are:

Source - implemented by default for &str and &[u8], used by the Lexer.

Slice - slices of Source, returned by Lexer::slice.

Traits

Chunk

A fixed, statically sized chunk of data that can be read from the Source.

Source

Trait for types the Lexer can read from.