Trait logos::source::Chunk

pub trait Chunk<'source>: Sized + Copy + PartialEq + Eq {
    const SIZE: usize;

    unsafe fn from_ptr(ptr: *const u8) -> Self;
}

A fixed, statically sized chunk of data that can be read from a Source.

This is implemented for u8, as well as borrowed byte arrays of any size.

Associated Constants

const SIZE: usize

The size of the chunk, in bytes.

Required methods

unsafe fn from_ptr(ptr: *const u8) -> Self

Create a chunk from a raw pointer.

Safety

The caller must ensure that ptr is well-aligned, non-dangling and properly initialised for a value of type Self. Additionally, the caller must ensure that ptr is valid for reads of at least Self::SIZE bytes.

Violating any of these conditions is immediate undefined behaviour.