Trait logos::LexerExt

pub trait LexerExt<'source> {
    type Token: Logos<'source>;
    fn as_lexer(&self) -> &Lexer<'source, Self::Token>;

    fn as_lexer_mut(&mut self) -> &mut Lexer<'source, Self::Token>;

    fn into_lexer(self) -> Lexer<'source, Self::Token>;

    fn span(&self) -> Span { ... }

    fn source(&self) -> &'source <Self::Token as Logos<'source>>::Source { ... }

    fn slice(
        &self
    ) -> &'source <<Self::Token as Logos<'source>>::Source as Source>::Slice { ... }

    fn remainder(
        &self
    ) -> &'source <<Self::Token as Logos<'source>>::Source as Source>::Slice { ... }

    fn extras<'lexer>(
        &'lexer self
    ) -> &'lexer <Self::Token as Logos<'source>>::Extras
    where
        'source: 'lexer,
    { ... }

    fn extras_mut<'lexer>(
        &'lexer mut self
    ) -> &'lexer mut <Self::Token as Logos<'source>>::Extras
    where
        'source: 'lexer,
    { ... }

    fn map_with_lexer<F, O>(self, op: F) -> MapWithLexer<'source, Self, F>
    where
        Self: Sized + Iterator,
        F: FnMut(Self::Item, &Lexer<'source, Self::Token>) -> O,
    { ... }

    fn boxed(self) -> BoxedLexer<'source, Self::Token, Self::Item>
    where
        Self: Sized + Iterator + 'source,
    { ... }

    fn lookahead(self) -> Lookahead<'source, Self>
    where
        Self: Sized + Iterator + 'source,
    { ... }
}

Extension methods for types that contain a Lexer.

Associated Types

type Token: Logos<'source>

The token type that the Lexer uses.

Required methods

fn as_lexer(&self) -> &Lexer<'source, Self::Token>

Get a reference to the underlying Lexer.

fn as_lexer_mut(&mut self) -> &mut Lexer<'source, Self::Token>

Get a mutable reference to the underlying Lexer.

fn into_lexer(self) -> Lexer<'source, Self::Token>

Consume self, and return the underlying Lexer.

Provided methods

fn span(&self) -> Span

The source position of the current token. This is equivalent to Lexer::span.

fn source(&self) -> &'source <Self::Token as Logos<'source>>::Source

The source that tokens are being read from. The return type of this method is determined by Logos::Source, and will be &str for most lexers. This is equivalent to Lexer::source.

fn slice(
&self
) -> &'source <<Self::Token as Logos<'source>>::Source as Source>::Slice

A slice containing the current token. The return type of this method is determined by Logos::Source, and will be &str for most lexers. This is equivalent to Lexer::slice.

fn remainder(
&self
) -> &'source <<Self::Token as Logos<'source>>::Source as Source>::Slice

A slice containing the remaining source. This is similar to Lexer::source, but starts at the end of the current token. The return type of this method is determined by Logos::Source, and will be &str for most lexers. This is equivalent to Lexer::remainder.

fn extras<'lexer>(
&'lexer self
) -> &'lexer <Self::Token as Logos<'source>>::Extras where
'source: 'lexer,

Get a reference to the lexer’s extras. This is a shorthand for calling LexerExt::as_lexer and accessing the extras field.

fn extras_mut<'lexer>(
&'lexer mut self
) -> &'lexer mut <Self::Token as Logos<'source>>::Extras where
'source: 'lexer,

Get a mutable reference to the lexer’s extras. This is a shorthand for calling LexerExt::as_lexer_mut and accessing the extras field.

fn map_with_lexer<F, O>(self, op: F) -> MapWithLexer<'source, Self, F> where
Self: Sized + Iterator,
F: FnMut(Self::Item, &Lexer<'source, Self::Token>) -> O,

Wrap the lexer in an Iterator that maps each token to another value, making use of the lexer in the process.

The returned iterator produces values by calling op for each token, passing both the token and a reference to the lexer as arguments.

See also Lexer::spanned, which uses this method to pair tokens with their source positions.

fn boxed(self) -> BoxedLexer<'source, Self::Token, Self::Item> where
Self: Sized + Iterator + 'source,

Box the lexer, returning a type-erased BoxedLexer.

This incurs a small performance penalty from dynamic dispatch, but makes it possible to name the type of the lexer when you might otherwise be unable to.

fn lookahead(self) -> Lookahead<'source, Self> where
Self: Sized + Iterator + 'source,

Wrap the Lexer in an Iterator that can use the peek and peek_mut methods to see the future.