Crate logos

Logos

Create ridiculously fast lexers

Logos has two goals:

To make it easy for you to create a lexer, so that you can spend your time on more complex problems.

To make that lexer faster than anything you’d write by hand.

To achieve this, Logos:

Combines all token definitions into a single deterministic state machine.

Optimizes branches into lookup tables or jump tables.

Prevents backtracking inside token definitions.

Unwinds loops, and batches reads to minimize bounds checking.

Does all of that heavy lifting at compile time!

Example

use logos::Logos;

#[derive(Logos, Debug, PartialEq)]
enum Token {
    // Tokens can be literal strings, of any length.
    #[token("fast")]
    Fast,

    #[token(".")]
    Period,

    // Or regular expressions.
    #[regex("[a-zA-Z]+")]
    Text,

    // We can also use this variant to define whitespace,
    // or any other matches we wish to skip.
    #[regex(r"[ \t\n\f]+", logos::skip)]
    Whitespace,
}

fn main() {
    let mut lex = Token::lexer("Create ridiculously fast lexers.");

    assert_eq!(lex.next(), Some(Ok(Token::Text)));
    assert_eq!(lex.span(), 0..6);
    assert_eq!(lex.slice(), "Create");

    assert_eq!(lex.next(), Some(Ok(Token::Text)));
    assert_eq!(lex.span(), 7..19);
    assert_eq!(lex.slice(), "ridiculously");

    assert_eq!(lex.next(), Some(Ok(Token::Fast)));
    assert_eq!(lex.span(), 20..24);
    assert_eq!(lex.slice(), "fast");

    assert_eq!(lex.next(), Some(Ok(Token::Text)));
    assert_eq!(lex.span(), 25..31);
    assert_eq!(lex.slice(), "lexers");

    assert_eq!(lex.next(), Some(Ok(Token::Period)));
    assert_eq!(lex.span(), 31..32);
    assert_eq!(lex.slice(), ".");

    assert_eq!(lex.next(), None);
}

Callbacks

Logos can also call arbitrary functions whenever a pattern is matched, which can be used to put data into a variant or perform more complicated lexing if regular expressions are too limiting.

use logos::{Logos, Lexer};

fn kilo(lex: &mut Lexer<Token>) -> Option<u64> {
    let slice = lex.slice();
    let n: u64 = slice[..slice.len() - 1].parse().ok()?; // skip 'k'
     
    Some(n * 1_000)
}

fn mega(lex: &mut Lexer<Token>) -> Option<u64> {
    let slice = lex.slice();
    let n: u64 = slice[..slice.len() - 1].parse().ok()?; // skip 'm'

    Some(n * 1_000_000)
}

#[derive(Logos, Debug, PartialEq)]
enum Token {
    #[regex(r"[ \t\n\f]+", logos::skip)]
    Whitespace,

    // Callbacks can use closure syntax, or refer
    // to a function defined elsewhere.
    //
    // Each pattern can have its own callback.
    #[regex("[0-9]+", |lex| lex.slice().parse().ok())]
    #[regex("[0-9]+k", kilo)]
    #[regex("[0-9]+m", mega)]
    Number(u64),
}

fn main() {
    let mut lex = Token::lexer("5 42k 75m");

    assert_eq!(lex.next(), Some(Ok(Token::Number(5))));
    assert_eq!(lex.slice(), "5");

    assert_eq!(lex.next(), Some(Ok(Token::Number(42_000))));
    assert_eq!(lex.slice(), "42k");

    assert_eq!(lex.next(), Some(Ok(Token::Number(75_000_000))));
    assert_eq!(lex.slice(), "75m");

    assert_eq!(lex.next(), None);
}

Callbacks can return any value implementing the CallbackResult trait. This trait is implemented already for Option, Result, bool and many more - see the documentation on callbacks for more information.

Token disambiguation

The rule of thumb is:

Longer beats shorter.

Specific beats generic.

If any two patterns could match the same input, such as fast and [a-zA-Z]+ from the example above, Logos will choose the longer and more specific pattern. In this case, that means that a string like “fast” will result in Token::Fast instead of Token::Text.

This is done by comparing the numeric priority of each definition. This priority is calculated using a simple set of rules:

A regex character class (like [a-z]) has a priority of 1.

A byte of text has a priority of 2.

The priority of a repetition or alternation is the priority of its shortest possible match. More specifically,

The shortest possible match for patterns like term* and term? is an empty string.

The shortest possible match for a pattern like term+ is just term, since + matches at least once.

The shortest possible match for an alternation is always the shortest alternative.

For example,

[a-zA-Z]+ has a priority of 1; the priority of [a-zA-Z] is 1, and the + quantifier doesn’t affect the shortest possible match.

foobar has a priority of 12; it’s 6 bytes, and each byte has a priority of 2.

(foo|hello)(bar)? has a priority of 6. foo is the shortest option in the alternation, with a priority of 6. The shortest possible match for (bar)? is an empty string, so it has a priority of 0.

If two definitions have the same priority and might match the same input, you’ll receive a compile error! Logos will point out the problematic definitions, and ask you to manually specify a priority to fix the issue.

For example, the patterns [abc]+ and [cde]+ both have a priority of 1, and will both match repeated cs. You can fix this ambiguity by changing the definition to #[regex("[abc]+", priority = 2)] within the derive macro. In this case, the definition using "[abc]+" now has a higher priority, so all sequences of c will match the [abc]+ definition instead.

How fast?

Ridiculously fast!

test identifiers                        ... bench:         647 ns/iter (+/- 27) = 1204 MB/s
test keywords_operators_and_punctuators ... bench:       2,054 ns/iter (+/- 78) = 1037 MB/s
test strings                            ... bench:         553 ns/iter (+/- 34) = 1575 MB/s

What about the changes you made?

I ran benchmarks on both this fork and the upstream repository, and rest assured that even after my changes, Logos is still just as fast. Hooray!

Acknowledgements

Pedrors for the Logos logo.

Thank you

Logos is very much a labor of love, so consider buying the real author a coffee. ☕

License

Logos is licensed under the terms of both the MIT license and the Apache License (Version 2.0), at your option.

See LICENSE-APACHE and LICENSE-MIT for details.

Re-exports

pub use crate::error::Error;

pub use crate::error::UnknownToken;

pub use crate::source::Source;

Modules

callback

Types and traits used within lexer callbacks.

error

Types and traits for error handling in Logos.

iter

Tools for working with lexers as iterators.

source

Traits for reading from different input sources.

Structs

Lexer

A Lexer allows you to read through a source (a type implementing the Source trait, like a string slice) and produce tokens using the Logos trait. It’s important to note that you should not implement Logos yourself, and should always use the derive macro instead. See the trait’s documentation for more details.

Skip

Used within callbacks to instruct the lexer to skip a token match.

Enums

Filter

A type that can be used within callbacks to either produce a field for a token or skip a token match.

Traits

LexerExt

Extension methods for types that contain a Lexer.

Logos

Types that can be lexed using Lexer.

Functions

skip

A predefined callback that unconditionally skips a token match.

Type Definitions

Span

A byte range in the source.

Derive Macros

Logos