Trait logos::Logos

pub trait Logos<'source>: Sized {
    type Extras;
    type Source: Source + ?Sized + 'source;
    type Error: Error<'source, Self>;
    fn lex(lexer: &mut Lexer<'source, Self>);

    fn lexer(source: &'source Self::Source) -> Lexer<'source, Self>
    where
        Self::Extras: Default,
    { ... }

    fn lexer_with_extras(
        source: &'source Self::Source, 
        extras: Self::Extras
    ) -> Lexer<'source, Self> { ... }
}

Types that can be lexed using Lexer.

The section below is quite long, so here’s a handy link to jump to the rest of the trait documentation.

Using the derive macro

The main way to interact with this crate is through #[derive(Logos)] - the Logos derive macro. The derive macro is responsible for generating an implementation of this trait, and it’s what does all the heavy lifting for you. A lexer would be useless if you couldn’t tell it what and how to lex, so the derive macro also accepts a healthy set of options to configure the generated lexer.

Attributes

The derive macro uses attributes to customize the generated Logos implementation. For the purposes of documentation, these are split into two categories:

Enum attributes, which are applied to an enum definition

Variant attributes, which are applied to an individual enum variant

It’s important to note that a single enum or enum variant can have multiple attributes applied to it.

Enum attributes

#[logos(error = SomeType)]

Sets the error type for this Logos implementation.

Logos will use this type to report lexing errors - namely, encountering an unknown token - but you can also use this type within callbacks to emit more detailed errors.

The error type is UnknownToken by default.

See the documentation on callbacks for details not covered here.

#[logos(extras = SomeType)]

Sets the extras type for this Logos implementation.

The Lexer will store a value of this type in its extras field, and you’re free to modify it as you wish. Because callbacks are passed &mut Lexer as an argument, this value can also be accessed and modified within callbacks, so it’s useful if you’d like to store additional state within your lexer.

The extras type is () by default.

See the documentation on callbacks, the extras type and the Lexer type for details not covered here.

#[logos(type T = SomeType)]

Specify the concrete type to use for the type parameter T.

If your lexer contains generic type parameters, Logos will emit an error - it doesn’t know what type should be used to fill in those type parameters. As an example, the following snippet fails to compile:

use logos::{Logos, Lexer};

// A magical type!
#[derive(Default)]
struct SuperMagic;

// A callback that produces a default value for `T`.
fn make_magic<'source, T>(lexer: &mut Lexer<'source, Token<T>>) -> T
where
    T: Default + 'source,
    Token<T>: Logos<'source>
{
    T::default()
}

#[derive(Logos)]
enum Token<T: Default> {
    #[token("magic", make_magic)]
    Magic(T)
}

However, we can use the type option to tell Logos what type T (or any other type parameter) should be:

#[derive(Logos)]
#[logos(type T = SuperMagic)] // This is important!
enum Token<T: Default> {
    #[token("magic", make_magic)]
    Magic(T)
}

At present, the derive macro does not perform generic implementations of the Logos trait, so you must always specify replacements for type parameters. This is likely to change in the future.

#[logos(subpattern NAME = "...")]

Define a subpattern named NAME that can be used within regular expressions.

Subpatterns allow you to avoid code duplication by factoring out the common parts of a regular expression, and additionally help readability by making it possible to “name” a regular expression’s component parts.

Inside of regular expressions, subpatterns are used by wrapping ?& and a subpattern name within parenthesis. For example, the expression (?&IDENTIFIER) refers to a subpattern named IDENTIFIER.

Variant attributes

#[token(...)] and #[regex(...)]

The #[token(...)] and #[regex(...)] attributes are used to add rules to your lexer, so that you can produce tokens based on certain criteria. Both of these attributes are very similar (and share common configuration options) so they’re grouped here for easier reading.

#[token("LITERAL", ...)]

Produce this variant when the provided literal string is found.

A #[token(...)] definition should be used for tokens that are represented solely by a specific piece of input, such as keywords or symbols. For example,

use logos::Logos;

#[derive(Logos)]
enum Token {
    #[token("if")]
    If,
    #[token("then")]
    Then,
    #[token("else")]
    Else,
    // ... and so on. You can use any string you'd like!
}

For tokens that may be represented by any input that fits a certain criteria (such as string literals or identifiers) you should use a #[regex(...)] definition instead.

#[regex("PATTERN", ...)]

Produce this variant when the provided regular expression matches.

A #[regex(...)] definition should be used for tokens that are represented by any input that fits a certain criteria - such as integer literals, identifiers, and similar. Integer literals could be nearly any sequence of the numbers 0 through 9, and writing out every possible number using #[token(...)] definitions would be a colossal waste of time.

We can use a #[regex(...)] definition for this instead, like so:

use logos::Logos;

#[derive(Logos, Debug, PartialEq)]
enum Token {
    // Callbacks are explained further below, but for now all you need to know is that we use the `logos::skip`
    // callback here so that we can ignore whitespace.
    #[regex(r"\s", logos::skip)]
    Whitespace,

    #[regex(r"[1-9]+[0-9]*")]
    Integer,
    // ... and so on. See below for more details on what you can and can't do.
}

let mut lexer = Token::lexer("89 144 233");

lexer.next();
assert_eq!(lexer.slice(), "89");

lexer.next();
assert_eq!(lexer.slice(), "144");

lexer.next();
assert_eq!(lexer.slice(), "233");

// We've reached the end of our input, so calling `next` again returns `None`
assert_eq!(lexer.next(), None);

The regular expression used within a #[regex(...)] definition is subject to some limitations. Most notably:

Look-around is not supported

Backreferences are not supported

Line anchors may not be used

Capture groups cannot be used to extract portions of the matched input.

If you’d like to perform more complicated lexing, you can use lexer callbacks, which are described below and in the documentation on callbacks.

Callbacks

Callbacks may be attached to a #[token(...)] or #[regex(...)] definition, and are called whenever a match occurs. They can be used to put data into a variant, skip a token match, or to emit errors. Callbacks can also perform more complicated lexing if regular expressions are too limiting.

Callbacks may be provided as closures, or as paths to functions defined elsewhere. They are typically passed as positional arguments to the #[token(...)] or #[regex(...)] attributes, like so:

use logos::Logos;

#[derive(Logos, Debug, PartialEq)]
enum Token {
    // We can use the predefined `skip` callback to ignore whitespace
    #[regex(r"\s", logos::skip)]
    Whitespace,

    // ... and we can also use a closure to put data into a variant
    #[regex("[0-9]+", |lex| lex.slice().parse().ok())]
    Integer(usize)
}

let mut lexer = Token::lexer("2000 305 8");

assert_eq!(lexer.next(), Some(Ok(Token::Integer(2000))));
// The whitespace between `2000` and `305` is skipped
assert_eq!(lexer.next(), Some(Ok(Token::Integer(305))));
// The whitespace between `305` and `8` is skipped
assert_eq!(lexer.next(), Some(Ok(Token::Integer(8))));
// We've reached the end of our input, so the lexer returns `None`
assert_eq!(lexer.next(), None);

Callbacks may also be passed using the callback option, which is described further in the section below.

See the documentation on callbacks for details not covered here.

Other options

The #[token(...)] and #[regex(...)] attributes also support additional configuration options.

Additional options must come after the pattern and callback. For example, the following is invalid:

use logos::Logos;

#[derive(Logos)]
enum Token {
    #[token(priority = 20, "fast")]
    Fast
}

priority = ...

Sets the priority of the definition. This argument’s value should be an integer.

The priority value is used to disambiguate tokens, and is normally calculated automatically. Logos will issue a compiler error if any two definitions have the same priority and could match the same input, so this option can be used to solve conflicts in that scenario.

For example, the following two definitions are in conflict:

use logos::Logos;

#[derive(Logos)]
enum Token {
    #[token("fast")]
    Fast,

    #[regex("(ridiculously)?fast(er|est)?")]
    RidiculouslyFast,
}

Both the #[regex("(ridiculously)?fast[er|est]?")] and #[token("fast")] definitions could match the input “fast”, and both have the same priority.

When you encounter a conflict like this, Logos will also use the compile error to suggest a priority value that isn’t ambiguous. In this case, fast has a priority of 8, so Logos suggested that we use a priority of 9 to disambiguate:

error: A definition of variant `RidiculouslyFast` can match the same input as another definition of variant `Fast`.

       hint: Consider giving one definition a higher priority: #[regex(..., priority = 9)]
  -> src\example.rs:8:13
  |
5 |     #[regex("fast[er|est]?")]
  |             ^^^^^^^^^^^^^^^

error: A definition of variant `Fast` can match the same input as another definition of variant `RidiculouslyFast`.

       hint: Consider giving one definition a higher priority: #[regex(..., priority = 9)]
 --> src\example.rs:8:13
  |
8 |     #[token("fast")]
  |             ^^^^^^

You can then solve this conflict by using the priority option:

use logos::Logos;

#[derive(Logos, Debug, PartialEq)]
enum Token {
    #[regex(r"\s", logos::skip)]
    Whitespace,

    #[token("fast", priority = 9)]
    Fast,

    #[regex("(ridiculously)?fast(er|est)?")]
    RidiculouslyFast,
}

let mut lexer = Token::lexer("fast faster fastest ridiculouslyfast");

assert_eq!(lexer.next(), Some(Ok(Token::Fast)));
assert_eq!(lexer.slice(), "fast");

assert_eq!(lexer.next(), Some(Ok(Token::RidiculouslyFast)));
assert_eq!(lexer.slice(), "faster");

assert_eq!(lexer.next(), Some(Ok(Token::RidiculouslyFast)));
assert_eq!(lexer.slice(), "fastest");

assert_eq!(lexer.next(), Some(Ok(Token::RidiculouslyFast)));
assert_eq!(lexer.slice(), "ridiculouslyfast");

// We've reached the end of our input, so calling `next` again returns `None`
assert_eq!(lexer.next(), None);

callback = ...

Sets the callback attached to this definition. This argument’s value should be a closure or a path to a function defined elsewhere.

For example,

use logos::Logos;

#[derive(Logos)]
enum Token {
    // We can use the predefined `skip` callback to ignore whitespace
    #[regex(r"\s", callback = logos::skip)]
    Whitespace,
}

Callbacks may also be passed as positional arguments, so you likely won’t find yourself needing to use the named form. See the section above and the documentation on callbacks for details not covered here.

ignore(...)

Ignore certain aspects of the input when performing comparisons. This argument’s value should be a parenthesized sequence of flags.

Valid flags are

case - Comparisons between characters will be entirely case-insensitive. This flag may not be used with ascii_case.

ascii_case - Comparisons between ASCII characters will be case-insensitive. This flag may not be used with case.

For example:

use logos::Logos;

#[derive(Logos, Debug, PartialEq)]
enum Token {
    #[regex(r"\s", logos::skip)]
    Whitespace,

    #[regex("[a-z]+[0-9a-z]*", ignore(ascii_case))]
    Identifier,
}

let mut lexer = Token::lexer("Case iS UNIMPORtant");

lexer.next();
assert_eq!(lexer.slice(), "Case");

lexer.next();
assert_eq!(lexer.slice(), "iS");

lexer.next();
assert_eq!(lexer.slice(), "UNIMPORtant");

// We've reached the end of our input, so calling `next` again returns `None`
assert_eq!(lexer.next(), None);

Associated Types

type Extras

The “extras” type, used to add state to a lexer.

Occasionally you’d like to add extra information to a lexer; perhaps for the sake of debugging, or just to handle more complicated input. Logos allows you to insert this extra information into a lexer using a concept of “extras” - a value stored in the lexer that you can access within callbacks and modify as you wish.

By default, Logos will just use the () type for extras. If you’d like to use a different type, you can use the derive macro’s extras option.

type Source: Source + ?Sized + 'source

The source type that tokens are lexed from.

This is str by default, but will be [u8] if you use a non-unicode byte string (or character class) within a #[token(...)] or #[regex(...)] definition.

You can use your own source type by passing a type implementing the Source trait to the derive macro’s source option.

type Error: Error<'source, Self>

The type used to report errors during lexing.

This is UnknownToken by default, but you can use your own error type by passing a type implementing the Error trait to the derive macro’s error option.

Required methods

fn lex(lexer: &mut Lexer<'source, Self>)

The heart of Logos.

This method is called during the lexing process, and is implemented by the logos-derive crate. As a reminder, you should never implement this trait yourself. Use the derive macro!

Provided methods

fn lexer(source: &'source Self::Source) -> Lexer<'source, Self> where
Self::Extras: Default,

Create a new Lexer for this token type.

fn lexer_with_extras(
source: &'source Self::Source,
extras: Self::Extras
) -> Lexer<'source, Self>

Create a new Lexer for this token type, using the provided extras value.

Note

In most cases, you can use Logos::lexer instead. You should only use this function if you need to set up your lexer in a way that doesn’t play nicely with the Default trait.